requirements.txt
sorl-thumbnail>=11.01
